toolkit/components/passwordmgr/test/test_bug_391514.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Login Manager</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>  
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: 391514
<p id="display"></p>
<div id="content" style="display: none">
  <!-- normal form. -->
  <form id="form1" action="formtest.js">
    <input  type="text"     name="uname">
    <input  type="password" name="pword">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <!-- limited username -->
  <form id="form2" action="formtest.js">
    <input  type="text"     name="uname" maxlength="4">
    <input  type="password" name="pword">
  </form>

  <!-- limited password -->
  <form id="form3" action="formtest.js">
    <input  type="text"     name="uname">
    <input  type="password" name="pword" maxlength="4">
  </form>

  <!-- limited username and password -->
  <form id="form4" action="formtest.js">
    <input  type="text"     name="uname" maxlength="4">
    <input  type="password" name="pword" maxlength="4">
  </form>


  <!-- limited username -->
  <form id="form5" action="formtest.js">
    <input  type="text"     name="uname" maxlength="0">
    <input  type="password" name="pword">
  </form>

  <!-- limited password -->
  <form id="form6" action="formtest.js">
    <input  type="text"     name="uname">
    <input  type="password" name="pword" maxlength="0">
  </form>

  <!-- limited username and password -->
  <form id="form7" action="formtest.js">
    <input  type="text"     name="uname" maxlength="0">
    <input  type="password" name="pword" maxlength="0">
  </form>


  <!-- limited, but ok, username -->
  <form id="form8" action="formtest.js">
    <input  type="text"     name="uname" maxlength="999">
    <input  type="password" name="pword">
  </form>

  <!-- limited, but ok, password -->
  <form id="form9" action="formtest.js">
    <input  type="text"     name="uname">
    <input  type="password" name="pword" maxlength="999">
  </form>

  <!-- limited, but ok, username and password -->
  <form id="form10" action="formtest.js">
    <input  type="text"     name="uname" maxlength="999">
    <input  type="password" name="pword" maxlength="999">
  </form>


  <!-- limited, but ok, username -->
  <!-- (note that filled values are exactly 8 characters) -->
  <form id="form11" action="formtest.js">
    <input  type="text"     name="uname" maxlength="8">
    <input  type="password" name="pword">
  </form>

  <!-- limited, but ok, password -->
  <!-- (note that filled values are exactly 8 characters) -->
  <form id="form12" action="formtest.js">
    <input  type="text"     name="uname">
    <input  type="password" name="pword" maxlength="8">
  </form>

  <!-- limited, but ok, username and password -->
  <!-- (note that filled values are exactly 8 characters) -->
  <form id="form13" action="formtest.js">
    <input  type="text"     name="uname" maxlength="8">
    <input  type="password" name="pword" maxlength="8">
  </form>

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/* Test for Login Manager: 391514 (Login Manager gets confused with
 * password/PIN on usaa.com)
 */

commonInit();

function startTest() {
  var i;

  is($_(1, "uname").value, "testuser", "Checking for filled username 1");
  is($_(1, "pword").value, "testpass", "Checking for filled password 1");

  for (i = 2; i < 8; i++) {
    is($_(i, "uname").value, "", "Checking for unfilled username " + i);
    is($_(i, "pword").value, "", "Checking for unfilled password " + i);
  }

  for (i = 8; i < 14; i++) {
    is($_(i, "uname").value, "testuser", "Checking for filled username " + i);
    is($_(i, "pword").value, "testpass", "Checking for filled password " + i);
  }

  // Note that tests 11-13 are limited to exactly the expected value.
  // Assert this lest someone change the login we're testing with.
  is($_(11, "uname").value.length, 8, "asserting test assumption is valid.");

  SimpleTest.finish();
}

window.addEventListener("runTests", startTest);

SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
